+++ src/themes/fennec/corners/corners.js
@@ -1,0 +1,7 @@
+export default {
+	data() {
+		return {
+			isRequested: new URL(window.location).searchParams.has('corners'),
+		}
+	},
+}