+++ src/themes/fennec/shell/shell.js
@@ -1,3 +1,4 @@
+import Corners from '/hud/corners/corners.vue'
 import FocusedPlayer from '/hud/focused-player/focused-player.vue'
 import PlayersAlive from '/hud/players-alive/players-alive.vue'
 import Radar from '/hud/radar/radar.vue'
@@ -8,6 +9,7 @@ import TopBar from '/hud/top-bar/top-bar.vue'
 
 export default {
 	components: {
+		Corners,
 		FocusedPlayer,
 		PlayersAlive,
 		Radar,