+++ changelog.md
@@ -7,6 +7,7 @@ This project does not adhere to Semantic Versioning.
 
 ## [Unreleased]
 ### Added
+* Added optional overlay corners to show that the HUD is active (enabled by default in the Electron overlay binary; can be enabled by appending `?corners`/`?transparent&corners` to the URL)
 * Added `preferences.isCsgo` option on the config page (theme `raw`) for backwards-incompatible changes to CS:GO
 
 ### Fixed